+++ pkg/identityserver/identityserver_test.go
@@ -305,6 +305,12 @@ func getIdentityServer(t *testing.T) (*IdentityServer, *grpc.ClientConn) {
 				Level: log.DebugLevel,
 			},
 		},
+		KeyVault: config.KeyVault{
+			Provider: "static",
+			Static: map[string][]byte{
+				"is-test": {0x00, 0x01, 0x02, 0x03, 0x04, 0x05, 0x06, 0x07, 0x08, 0x09, 0x0A, 0x0B, 0x0C, 0x0D, 0x0E, 0x0F},
+			},
+		},
 	}})
 	conf := &Config{
 		DatabaseURI: dbConnString,