+++ pkg/identityserver/store/invitation_store_test.go
@@ -30,7 +30,7 @@ func TestInvitationStore(t *testing.T) {
 	a := assertions.New(t)
 	ctx := test.Context()
 
-	now := time.Now()
+	now := cleanTime(time.Now())
 
 	WithDB(t, func(t *testing.T, db *gorm.DB) {
 		prepareTest(db, &Account{}, &User{}, &Invitation{})